+++ components/BoardPostContainer/style.module.scss
@@ -2,3 +2,7 @@
   margin-top: 24px;
   margin-bottom: 150px;
 }
+
+.spinner {
+  margin: 100px auto;
+}
