+++ components/DropDown/index.tsx
@@ -2,6 +2,8 @@ import React, { useState } from 'react';
 import { SortType, SORT_TYPE_DICT } from '@/constants/sortOption';
 import SortIcon from '@/public/svgs/sort-icon.svg';
 import style from './style.module.scss';
+import useIsMobile from '@/hooks/useIsMobile';
+import cn from 'classnames';
 
 interface OptionProps {
   label: SortType;
@@ -13,16 +15,23 @@ interface DropdownProps {
 }
 
 const DropDown = ({ options, handleClickItem }: DropdownProps) => {
+  const isMobile = useIsMobile();
   const [isVisible, setIsVisible] = useState(false);
+  const [labelName, setLabelName] = useState<SortType>('recent');
 
   const onClickItem = (label: SortType) => {
     setIsVisible(false);
+    setLabelName(label);
     handleClickItem(label);
   };
 
   return (
     <div className={style.wrapper}>
-      <button className={style.button} onClick={() => setIsVisible(!isVisible)}>
+      <button
+        className={cn(style.button, { [style.mobile]: isMobile })}
+        onClick={() => setIsVisible(!isVisible)}
+      >
+        {!isMobile && <span>{SORT_TYPE_DICT[labelName]}</span>}
         <SortIcon />
       </button>
