+++ components/DropDown/style.module.scss
@@ -7,7 +7,7 @@
 .button {
   cursor: pointer;
 
-  width: 42px;
+  width: 130px;
   height: 42px;
 
   border-radius: 12px;
@@ -18,6 +18,18 @@
   &:hover {
     background-color: $cool-gray200;
   }
+
+  display: flex;
+  align-items: center;
+  justify-content: space-around;
+
+  span {
+    @include font-base($cool-gray800, 400, 16px, 24px);
+  }
+}
+
+.mobile {
+  width: 42px;
 }
 
 .container {
