+++ hooks/useIsMobile.ts
@@ -1,0 +1,24 @@
+import { useState, useEffect } from 'react';
+import { MOBILE_SIZE } from '@/constants/windowSize';
+
+const useIsMobile = (): boolean => {
+  const [isMobile, setIsMobile] = useState<boolean>(false);
+
+  useEffect(() => {
+    const handleResize = () => {
+      setIsMobile(window.innerWidth <= MOBILE_SIZE);
+    };
+
+    handleResize();
+
+    window.addEventListener('resize', handleResize);
+
+    return () => {
+      window.removeEventListener('resize', handleResize);
+    };
+  }, []);
+
+  return isMobile;
+};
+
+export default useIsMobile;
